refactor(encoding): Phase 1e Stage A+C batch 15 — lift optimal-plan emit + finish
Migrate `emit_optimal_plan` onto `impl MatchTable` (pure storage helper
over `window` + `offset_hist`) and `finish_optimal_plan` onto `impl
BtMatcher` (restores the seven `opt_*_scratch` buffers the parser
borrowed via `core::mem::take`).

Callers updated: the six `\$self.finish_optimal_plan(...)` sites inside
`build_optimal_plan_impl!` switch to `\$self.bt.finish_optimal_plan(...)`,
and the single `self.emit_optimal_plan(...)` driver call site switches
to `self.table.emit_optimal_plan(...)`.

Build (default + --no-default-features) + strict clippy + 381 lib
tests (incl. level22 ratio gate) green.
